feat: implemented mergesort; update: changed type to double; todo: refactor mergesort

+++ misc/swap.h
@@ -1,7 +1,7 @@
 #ifndef SWAP_H
 #define SWAP_H
 
-int swap (int* elem1, int* elem2) {
+int swap (double* elem1, double* elem2) {
 	int temp = *elem1;
 	*elem1 = *elem2;
 	*elem2 = temp;
